+++ packages/components/list/src/ListItem/ListItem.tsx
@@ -37,3 +37,5 @@ export function ListItem({
     </li>
   );
 }
+
+ListItem.displayName = 'ListItem';