+++ packages/components/notification/src/NotificationItem/NotificationItemContainer.tsx
@@ -22,6 +22,7 @@ export class NotificationItemContainer extends Component<
   NotificationItemContainerState
 > {
   static defaultProps = defaultProps;
+  displayName: 'NotificationItemContainer';
 
   timer: number | null = null;