+++ packages/components/card/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.3.7](https://github.com/contentful/forma-36/compare/@contentful/f36-card@4.3.6...@contentful/f36-card@4.3.7) (2022-03-10)
+
+**Note:** Version bump only for package @contentful/f36-card
+
+
+
+
+
+## [4.3.6](https://github.com/contentful/forma-36/compare/@contentful/f36-card@4.3.5...@contentful/f36-card@4.3.6) (2022-03-09)
+
+**Note:** Version bump only for package @contentful/f36-card
+
+
+
+
+
+## [4.3.5](https://github.com/contentful/forma-36/compare/@contentful/f36-card@4.3.4...@contentful/f36-card@4.3.5) (2022-03-07)
+
+
+### Bug Fixes
+
+* fix entry card description ([#1930](https://github.com/contentful/forma-36/issues/1930)) ([4c99709](https://github.com/contentful/forma-36/commit/4c997090dcc131b36a9004bdf40bc53dc744bc27))
+
+
+
+
+
+## [4.3.4](https://github.com/contentful/forma-36/compare/@contentful/f36-card@4.3.3...@contentful/f36-card@4.3.4) (2022-03-03)
+
+
+### Bug Fixes
+
+* migrate links from master to main  ([#1923](https://github.com/contentful/forma-36/issues/1923)) ([607301d](https://github.com/contentful/forma-36/commit/607301d57a2e83190d2aa298120ddb8493e8c429))
+
+
+
+
+
 ## [4.3.3](https://github.com/contentful/forma-36/compare/@contentful/f36-card@4.3.2...@contentful/f36-card@4.3.3) (2022-03-02)
 
 **Note:** Version bump only for package @contentful/f36-card