+++ packages/components/note/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.17](https://github.com/contentful/forma-36/compare/@contentful/f36-note@4.2.16...@contentful/f36-note@4.2.17) (2022-03-10)
+
+**Note:** Version bump only for package @contentful/f36-note
+
+
+
+
+
+## [4.2.16](https://github.com/contentful/forma-36/compare/@contentful/f36-note@4.2.15...@contentful/f36-note@4.2.16) (2022-03-09)
+
+**Note:** Version bump only for package @contentful/f36-note
+
+
+
+
+
+## [4.2.15](https://github.com/contentful/forma-36/compare/@contentful/f36-note@4.2.14...@contentful/f36-note@4.2.15) (2022-03-03)
+
+
+### Bug Fixes
+
+* migrate links from master to main  ([#1923](https://github.com/contentful/forma-36/issues/1923)) ([607301d](https://github.com/contentful/forma-36/commit/607301d57a2e83190d2aa298120ddb8493e8c429))
+
+
+
+
+
 ## [4.2.14](https://github.com/contentful/forma-36/compare/@contentful/f36-note@4.2.13...@contentful/f36-note@4.2.14) (2022-03-02)
 
 **Note:** Version bump only for package @contentful/f36-note