+++ packages/core/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.3](https://github.com/contentful/forma-36/compare/@contentful/f36-core@4.2.2...@contentful/f36-core@4.2.3) (2022-03-10)
+
+
+### Bug Fixes
+
+* **global-styles:** remove min-height on body ([#1937](https://github.com/contentful/forma-36/issues/1937)) ([6e67741](https://github.com/contentful/forma-36/commit/6e677415131057def3db3b906fba593c200dc3f3))
+
+
+
+
+
+## [4.2.2](https://github.com/contentful/forma-36/compare/@contentful/f36-core@4.2.1...@contentful/f36-core@4.2.2) (2022-03-09)
+
+
+### Bug Fixes
+
+* grid examples ([#1933](https://github.com/contentful/forma-36/issues/1933)) ([8f16c38](https://github.com/contentful/forma-36/commit/8f16c38f0c4c0f6c12df7e854c9bc4c15fdfd930))
+
+
+
+
+
+## [4.2.1](https://github.com/contentful/forma-36/compare/@contentful/f36-core@4.2.0...@contentful/f36-core@4.2.1) (2022-03-03)
+
+
+### Bug Fixes
+
+* migrate links from master to main  ([#1923](https://github.com/contentful/forma-36/issues/1923)) ([607301d](https://github.com/contentful/forma-36/commit/607301d57a2e83190d2aa298120ddb8493e8c429))
+
+
+
+
+
 # [4.2.0](https://github.com/contentful/forma-36/compare/@contentful/f36-core@4.1.1...@contentful/f36-core@4.2.0) (2022-02-17)